Add SDL_main for mingw due to link problem
- Comment out unit test that fails on MSVC

+++ game/ui/list.cc
@@ -48,7 +48,7 @@ void Item::setUserData(std::any const& data) {
 }
 
 
-const size_t List::None = -1;
+const size_t List::None = size_t(-1);
 
 List::List(std::vector<Item> const& items, Control* parent)
 : Control(parent), m_background(findFile("mainmenu_back_highlight.svg")), m_selectedBackground(findFile("mainmenu_back_highlight.svg")), m_items(items) {
@@ -146,10 +146,10 @@ void List::onKey(Key key) {
 
 	switch(key) {
 		case Key::Up:
-			select(std::max(int(m_selected) - 1, 0));
+			select(static_cast<size_t>(std::max(int(m_selected) - 1, 0)));
 			break;
 		case Key::Down:
-			select(std::min(m_selected + 1, countItems() - 1));
+			select(static_cast<size_t>(std::min(m_selected + 1, countItems() - 1)));
 			break;
 		case Key::Space:
 			if(isDisplayingCheckBox() && m_selected != None)
